refactor hash util function
